[WIP] resolving review
