+++ locationsyncservice/src/main/java/com/shuttl/location_pings/callbacks/LocationPingServiceCallback.kt
@@ -1,0 +1,10 @@
+package com.shuttl.location_pings.callbacks
+
+import com.shuttl.location_pings.data.model.entity.GPSLocation
+
+interface LocationPingServiceCallback {
+    fun afterSyncLocations(locations: List<GPSLocation>?)
+    fun errorWhileSyncLocations(error: String?)
+    fun serviceStarted()
+    fun serviceStopped()
+}
