+++ locationsyncservice/src/main/java/com/shuttl/location_pings/data/repo/LocationRepo.kt
@@ -2,11 +2,11 @@ package com.shuttl.location_pings.data.repo
 
 import android.text.TextUtils
 import android.util.Log
+import com.shuttl.location_pings.callbacks.LocationPingServiceCallback
 import com.shuttl.location_pings.config.components.LocationRetrofit
 import com.shuttl.location_pings.data.dao.GPSLocationsDao
 import com.shuttl.location_pings.data.model.entity.GPSLocation
 import com.shuttl.location_pings.data.model.request.SendLocationRequestBody
-import com.shuttl.location_pings.service.LocationPingService
 import kotlinx.coroutines.*
 
 class LocationRepo(private val locationsDao: GPSLocationsDao?) {
@@ -19,7 +19,6 @@ class LocationRepo(private val locationsDao: GPSLocationsDao?) {
             locationsDao?.deleteOldestLocation()
         }
         locationsDao?.addLocation(location)
-        printLocations()
     }
 
     fun clearLocations() = GlobalScope.async(Dispatchers.IO) {
@@ -31,15 +30,13 @@ class LocationRepo(private val locationsDao: GPSLocationsDao?) {
                       userId: String = "",
                       bookingId: String = "",
                       batchSize: Int,
-                      callback: LocationPingService.LocationPingServiceCallback?) {
+                      callback: LocationPingServiceCallback?) {
         GlobalScope.launch(Dispatchers.IO) {
             val locations = locationsDao?.getLimitedLocations(batchSize)
             if (locations?.isNotEmpty() == true) {
                 try {
                     if (TextUtils.isEmpty(url)) {
                         Log.e(TAG, "No Url Found")
-                    } else {
-                        Log.i(TAG, "url: $url")
                     }
                     val response = LocationRetrofit.locationAPI.syncLocation(
                         url,
@@ -58,10 +55,6 @@ class LocationRepo(private val locationsDao: GPSLocationsDao?) {
 
             }
         }
-    }
-
-    fun printLocations() {
-        Log.i(TAG, locationsDao?.locations().toString())
     }
 
     fun deleteEntries(timeStamp: String) {
